Implement dumb "profiling" of serial/parallel filters
Add two stages to example pipeline, modifying filter so that it has
a name and is passed an integer parameter to indicate how long to sleep
for. This is just a dumb way of seeing how tokens get passed through a
  sequence of serial/parallel filters.

It needs more work for me to fully understand the behaviour but it
appears that a full serial pipeline will still run stages in parallel,
just that there's no more than one token passing through any given
filter at once.