Check number of filter instances in tbb pipeline
Whilst TBB pipeline filters may run in parallel, the concurrency
appears to be on the level of concurrent method calls on single
filter instances.

Add reporting in IntMultiplier to output value of this pointer,
and add a static data member to record number of instances in
use. This confirms that only one instance of a given filter is
in use, with concurrent calls to the processing method.