Add basic example of TBB bounded queue
Implement basic (and rather dumb) producer/consumer system using
concurrent_bounded_queue. Share queue between a producer object and
a couple of consumers. Start each in a tbb_thread call, which is not
perfect but is good for simplicity. Producer just enqueues, consumers
pop items as long as the queue isn't empty. As the producer is started
first, we should have non-empty queue for consumers, but guess this
isn't guaranteed.

TODO : look at techniques, if available, for handling this potential
order issue safely.