Access remote content-type with libcurl
Add example of using the libcurl C API to get the header of a
remote file and report the content-type entry.

Ad entry to README noting that this may not be reliable if the
server does not know the "real" type. In these cases it will report
as text or octet-stream.